Clean a bunch of lines of code and some others
Firstly, I cleaned the code and made 1.3 lines of code down to 400 to
make it much cleaner and optimized? I'm not sure if that'll be part of
it but hope it is.

Secondly, I added a favicon.ico to the website to make it cuter LOL.

Third, I swapped around some elements from div to figure and figcaption
so I can style them easier and better with CSS, took a lot of time to do
though. I can say it was worth it to make the code look cleaner.