Add a bunch of new features
Added the header that has its position set to stick, as well contents
inside the header such as the redirects for the webpage itself and
added a search button that'll be used in the future, as of now it has no
function.

Also changed some colors and style of some elements.