Change design, fix header and add footer
Firstly, I changed the design and color scheme of the elements on the
site, it's still far from perfect but we're getting there. The colors
are still overlapping and it isn't that nice to look at.

I also fixed the header and made it cleaner, added a footer as well but
no footer content as of now since I'm not sure what it is I'm
supposed to be adding there.